requirements.txt
flask
gunicorn
firebase-admin
python-telegram-bot
requests
pillow